// apps/frontend/types/openapi.ts
export interface components {
    schemas: {
        Error401: {
            message: string;
        };
        FolderDataItemMeta: {
            createdAt: number;
            updatedAt: number;
        };
        FolderDataItemBase: {
            name: string;
            path: string;
            _meta: components["schemas"]["FolderDataItemMeta"];
        };
        /** @enum {string} */
        ItemType: "FOLDER" | "FILE";
        FolderDataItemFileBase: components["schemas"]["FolderDataItemBase"] & {
            itemType: components["schemas"]["ItemType"] & "FILE";
            src: string;
        };
        /** @enum {string} */
        FileType: "IMAGE" | "VIDEO" | "AUDIO" | "UNKNOWN";
        FolderDataItemAudio: components["schemas"]["FolderDataItemFileBase"] & {
            fileType: components["schemas"]["FileType"] & "AUDIO";
            metadata: {
                album: string | null;
                bpm: number | null;
                year: number | null;
                artists: string[];
                bitrate: number | null;
                duration: number | null;
            };
        };
        FolderDataSize: {
            width: number;
            height: number;
        };
        FolderDataItemImage: components["schemas"]["FolderDataItemFileBase"] & {
            fileType: components["schemas"]["FileType"] & "IMAGE";
            metadata: components["schemas"]["FolderDataSize"];
        };
        FolderDataItemVideo: components["schemas"]["FolderDataItemFileBase"] & {
            fileType: components["schemas"]["FileType"] & "VIDEO";
            metadata: components["schemas"]["FolderDataSize"];
        };
        FolderDataItemUnknown: components["schemas"]["FolderDataItemFileBase"] & {
            fileType: components["schemas"]["FileType"] & "UNKNOWN";
        };
        FolderDataItemFile: components["schemas"]["FolderDataItemAudio"] | components["schemas"]["FolderDataItemImage"] | components["schemas"]["FolderDataItemVideo"] | components["schemas"]["FolderDataItemUnknown"];
        PostResponse: {
            _meta: {
                id: string;
                createdAt: number;
                updatedAt: number;
            };
            text: string;
            attachments: components["schemas"]["FolderDataItemFile"][];
        };
        PostsResponse: {
            _meta: {
                cursorPrevious: number | null;
                cursorNext: number | null;
            };
            rows: components["schemas"]["PostResponse"][];
        };
        Error400: {
            message: string;
        };
        PostCreateRequest: {
            files: string[];
            text: string;
        };
        Error404: {
            message: string;
        };
        PostUpdateRequest: {
            text: string;
            attachments: (components["schemas"]["FolderDataItemFile"] | null)[];
            files: string[];
        };
        FolderDataItemFolder: components["schemas"]["FolderDataItemBase"] & {
            itemType: components["schemas"]["ItemType"] & "FOLDER";
        };
        FolderDataResponse: {
            folders: components["schemas"]["FolderDataItemFolder"][];
            files: components["schemas"]["FolderDataItemFile"][];
            file: components["schemas"]["FolderDataItemFile"] | null;
            pathDirectory: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
// ...
type FlattenedDeepRequired<T> = {
    [K in keyof T]-?: FlattenedDeepRequired<T[K] extends unknown[] | undefined | null ? Extract<T[K], unknown[]>[number] : T[K]>;
};
type ReadonlyArray<T> = [
    Exclude<T, undefined>
] extends [
    unknown[]
] ? Readonly<Exclude<T, undefined>> : Readonly<Exclude<T, undefined>[]>;
export const itemTypeValues: ReadonlyArray<FlattenedDeepRequired<components>["schemas"]["ItemType"]> = ["FOLDER", "FILE"];
export const folderDataItemFileBaseItemTypeValues: ReadonlyArray<FlattenedDeepRequired<components>["schemas"]["FolderDataItemFileBase"]["itemType"]> = ["FILE"];
export const fileTypeValues: ReadonlyArray<FlattenedDeepRequired<components>["schemas"]["FileType"]> = ["IMAGE", "VIDEO", "AUDIO", "UNKNOWN"];